Allow the hot-loop-comparison unroller to handle some comparisons that are stable across loops.
This pass handles loops that look like

	someBound = ...
	for ix in sequence(a,b):
		if ix < someBound:
			innerExpr

by trying to optimize away the "ix < someBound" case. If possible we tranform the loop into something
like
	someBound = ...
	ix = a
	while ix < b:
		stableLoopSize = (someBound - ix)
		if stableLoopSize > 4:
			top = ix + stableLoopSize
			while ix < top:
				innerExpr
				ix = ix + 1
		else:
			if ix < someBound:
				innerExpr
			ix = ix + 1

This often happens when we're iterating over vectors, and the 'if' statements are boundschecks
that we're almost always inside.

In the old version of the code, we couldn't handle it if the comparison statements we're optimizing
away contained expressions that were stable within the loop and didn't depend on 'ix'.