Removing broken GPU support for 8-bit integers.
While PTX supports 8-bit registers and values, the destination registers
of most instructions are not allowed to be smaller than 16 bits. The
way the NativeToPTX conversion is implemented, it is not trivial to
support 8-bit integers, so I disabled them. The same had been done
with 16-bit floats which have the same issue (most instructions write
32-bit floating-point results).