Adds initial support for running code that throws on GPU.
An index vector initialized to 0xFF is passed to the CUDA kernel
which is updates it to the correct value (corresponding to a type;
same mechanism as the one used for multiple return types) if the
execution terminates without throwing an exception. If an exception
in thrown, the value remains 0xFF. The values of this index vector
are checked after running the CUDA kernel and if any are found to
be 0xFF, we return an empty vector for the CUDAVectorApply operation.
We will upgrade to returning the appropriate exception, by executing
on the CPU the function on the first input index that threw an exception.