Modifies 'convertForGPUExecution' so that it may be applied incrementally.
Prior to this modification, convertForGPUExecution had to be executed after
inlining and CFG optimization and variable renaming were performed, otherwise
it would leave the CFG in an invalid state. With this update, it may be applied
incrementally to each CFG right before it gets inlined into the root CFG.
This allows to remove unsupported operations, and to replace vector operations
with GPU vector operations (not yet implemented).